+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2019-10-24
+
+- Added
+  - More deduplication, now also on the glTF accessor and mesh level
+  - Additional CLI options for ignoring mesh, line, or point geometry
+  - (experimental) serialization of glTF manifest into sqlite
+    - Can only be used when texture/buffer data is referenced and not embedded
+    - Potentially could be used for dynamically generating glTF variants with subsets of the original model
+    - Additional CLI option for serializing glTF manifest into sqlite
+    - Note that the schema of the sqlite database might change
+
 ## [0.6.4] - 2019-10-22
 
 - Added